+++ core/src/main/java/react4j/arez/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains the code to integrate with <a href="https://arez.github.io">Arez</a> reactive state library.
- */
-package react4j.arez;