+++ doc-examples/src/main/java/react4j/examples/arez/step1/Footer.java
@@ -1,9 +1,9 @@
 package react4j.examples.arez.step1;
 
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 import react4j.dom.proptypes.html.AnchorProps;
 import react4j.dom.proptypes.html.BtnProps;
 import react4j.dom.proptypes.html.HtmlProps;
@@ -11,9 +11,9 @@ import react4j.examples.arez.step1.model.AppData;
 import react4j.examples.arez.step1.model.FilterMode;
 import static react4j.dom.DOM.*;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class Footer
-  extends ReactArezComponent
+  extends Component
 {
   @Nullable
   @Override