+++ processor/src/main/java/react4j/processor/ComponentType.java
@@ -1,0 +1,12 @@
+package react4j.processor;
+
+/**
+ * Enum indicating type of component.
+ * This is a clone of @ReactComponent.Type and should be kept aligned
+ */
+enum ComponentType
+{
+  STATEFUL,
+  TRACKING,
+  MAYBE_TRACKING
+}