+++ processor/src/test/resources/bad_input/com/example/prop/CollectionArrayArezProp.java
@@ -2,14 +2,14 @@ package com.example.prop;
 
 import arez.annotations.ArezComponent;
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class CollectionArrayArezProp
-  extends ReactArezComponent
+  extends Component
 {
   @ArezComponent( allowEmpty = true )
   static abstract class AnArezElement