+++ processor/src/test/resources/input/com/example/arez/AutorunArezReactComponent.java
@@ -2,13 +2,13 @@ package com.example.arez;
 
 import arez.annotations.Observe;
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class AutorunArezReactComponent
-  extends ReactArezComponent
+  extends Component
 {
   @Observe
   void myAutorun()