+++ processor/src/test/resources/input/com/example/arez/ComponentFunctionalInterfaceProp.java
@@ -1,14 +1,14 @@
 package com.example.arez;
 
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class ComponentFunctionalInterfaceProp
-  extends ReactArezComponent
+  extends Component
 {
   @FunctionalInterface
   public interface TestFunction