+++ processor/src/test/resources/input/com/example/arez/ComponentJsFunctionProp.java
@@ -2,14 +2,14 @@ package com.example.arez;
 
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsFunction;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class ComponentJsFunctionProp
-  extends ReactArezComponent
+  extends Component
 {
   @JsFunction
   public interface TestFunction