+++ processor/src/test/resources/input/com/example/arez/ComponentShouldNotUpdateOnChangeProp.java
@@ -1,15 +1,15 @@
 package com.example.arez;
 
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.Feature;
 import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class ComponentShouldNotUpdateOnChangeProp
-  extends ReactArezComponent
+  extends Component
 {
   @Prop( shouldUpdateOnChange = Feature.DISABLE )
   protected abstract String getValue();