+++ processor/src/test/resources/input/com/example/arez/lifecycle/ActionOnLifecycleComponent.java
@@ -2,14 +2,14 @@ package com.example.arez.lifecycle;
 
 import javax.annotation.Nullable;
 import javax.xml.ws.Action;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.PostMount;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class ActionOnLifecycleComponent
-  extends ReactArezComponent
+  extends Component
 {
   @Nullable
   @Override