+++ processor/src/test/resources/input/com/example/arez/memoize/ParameterizedReturnMemoizeComponent.java
@@ -3,13 +3,13 @@ package com.example.arez.memoize;
 import arez.annotations.Memoize;
 import java.util.function.Consumer;
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class ParameterizedReturnMemoizeComponent
-  extends ReactArezComponent
+  extends Component
 {
   @Nullable
   @Override