+++ processor/src/test/resources/input/com/example/arez/memoize/PropAndMemoizeComponent.java
@@ -2,14 +2,14 @@ package com.example.arez.memoize;
 
 import arez.annotations.Memoize;
 import javax.annotation.Nullable;
+import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
-import react4j.arez.ReactArezComponent;
 
-@ReactComponent
+@ReactComponent( type = ReactComponent.Type.TRACKING )
 abstract class PropAndMemoizeComponent
-  extends ReactArezComponent
+  extends Component
 {
   @Prop
   protected abstract String getValue();